refactor: Move 'getStatus' to 'Auth'
